requirements.txt
numpy==1.26.4
pandas
scipy==1.14.0
ipykernel
yfinance
mysql-connector-python
sqlalchemy
praw
python-dotenv
vaderSentiment
scikit-learn
statsmodels==0.14.2
matplotlib
pmdarima
arch
openpyxl